requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
python-docx
fpdf2
duckduckgo-search